Add support for MacOS (#260)
* Use correct config folder on MacOS.

* Move -p before path on mkdir commands

For some reason, appending -p to end doesn't work on MacOS.

* Use open instead of xdg-open on mac os

open command is MacOS replacement of xdg-open.

* Surround some paths with quotation marks

The config path on mac has whitespace here: ...Application Support...
Which breaks many shell commands if not surrounded with quotation marks

* Seperate legendary binaries to folders and add MacOS binary

Q: How did you get this?
A:
- Clone legendary source from github
- Install pyinstaller using pip
- cd to 'legendary/legendary' and run 'pyinstaller cli.py --name legendary -F'

NOTE: Pyinstaller does not have cross compile support so you need to do this on MacOS.

* Use zsh on macOS instead of bash.

It fixes the 'legendary gets instantly killed when called from heroic' bug
Also bash is deprecated on MacOS and zsh is default
So it's better to use zsh anyways.

* 'Fix' stopping game installations on mac

It's not a perfect solution as it will stop all installations on MacOS.
But there is nothing we can do from our side.

As @LukeVader-IV said, Legendary's owner does not want support mac, so I just did this for now.

Maybe we can add a note to README?

* Create 'dist-mac' command for packaging for MacOS.