Change how legendary/gogdl is invoked
This changes every legendary/gogdl-related function to call
runLegendaryCommand/runGogdlCommand. With that, it unifies
the ~3 different approaches used before into one (hopefully)
easy-to-understand function.

This has the added benefit that if this function breaks, it'll
break everywhere. This should prevent a scenario like in the past
few days from happening, as either everything or nothing works.