Refactor Legendary/GOGDL handling (#1092)
* Revert "[General] Add Legendary and GOGDL binaries to env.PATH (#1081)"

This reverts commit 710b3677066b9ce7272f42fa7e508ffcb5ddf033.

* Run legendary/gogdl version check after updating settings

* Remove deprecated & unused API

* Change how legendary/gogdl is invoked

This changes every legendary/gogdl-related function to call
runLegendaryCommand/runGogdlCommand. With that, it unifies
the ~3 different approaches used before into one (hopefully)
easy-to-understand function.

This has the added benefit that if this function breaks, it'll
break everywhere. This should prevent a scenario like in the past
few days from happening, as either everything or nothing works.

* Fix progress logging on Linux

There were a few unnecessary newlines here that I remove with .trim()

* Update translations

* Fixup GOGDL name

* Change usage in import()

* Fix GOGDL

When using a shell, the quotes around paths got removed automatically.
Now with spawn, this no longer happened and apparently
GOGDL wasn't happy about it.

* De-clutter close event

Not sure why this was done like this, seems to still work fine now

* Close all legendary/gogdl processes on exit

This is kinda hacky, but it stops weird things from happening when
starting a download, quitting heroic, and then starting it again

* Add a new 'error' property to ExecResult

* Utilize ExecResult.error everywhere

This changes just about all implementations of ExecResult to use
the new error property

* Implement discussed changes

* Make home always point to /home + rename it to userHome