Add a new "PathSelectionBox" component
This adds a new component to be used in the Settings. It makes it easy
to request a file/directory path from the user.

Essentially, it is just a `TextInputField` with built-in icons, but we
use this kind of thing a lot in the codebase, so generalizing it seemed
like a good idea.