[Fix] EGL sync on macOS/Linux (#2501)
* Add a new "PathSelectionBox" component

This adds a new component to be used in the Settings. It makes it easy
to request a file/directory path from the user.

Essentially, it is just a `TextInputField` with built-in icons, but we
use this kind of thing a lot in the codebase, so generalizing it seemed
like a good idea.

* Make use of `PathSelectionBox` in EGL sync

This component was entirely broken on macOS/Linux. If a Wine prefix path
was selected, you could only "unsync", not "sync". Further, sync also
wasn't happening when the user selected the path.
Both of those issues combined meant that you could never actually sync
between EGL and Legendary/Heroic

I've now replaced this with a single path selector that automatically
syncs/un-syncs depending on the user selecting or clearing the path.
This seems more intuitive to me.

* Use `PathSelectionBox` everywhere else

* Only call `onPathChange` in the `onBlur` event

Saving after each keystroke seems unnecessary for paths

* Remove unused translation keys