Remove "launch --json" call
The purpose of this was making sure the user didn't modify `config.ini`
in a problematic way. Now that we alone control the config file, we
don't have to worry about this anymore.

As a side effect, this also resolves issues for MacOS users without a
CrossOver installation (Legendary would try to auto-detect CX & crash if
nothing is found)