Treat GPTK Wine like regular Wine
When launching GPTK, Heroic simply uses the underlying `wine64` binary.
This binary is a (mostly) normal Wine binary, meaning it's meant to be
launched with the WINEPREFIX environment variable set to the desired
Wineprefix & a single argument (the executable to run). `gameportingtoolkit`
on the other hand is the wrapper around Wine, not taking in
a WINEPREFIX & instead requiring two arguments (the prefix, then the
executable to run)
We were running `wine64`, but were passing it parameters intended
for `gameportingtoolkit`. I'm not sure how this ever worked

While touching the responsible function (getWineFlags) I've also made
it a little more type-safe & removed the unused `gameSettings` parameter