Add type definitions for all Legendary commands used in the project
This allows for stricter type checking (coming in the next commit).
The general idea behind this is:
- each command is represented by an object
- each of those objects has a `subcommand` property containing the name of the
  command, and other required properties for required parameters (for
  example, `launch` requires an `AppName`)
- optional parameters are represented by optional object properties with the
  only valid value of `true`

Base types (seen in `base.ts`) are general-purpose types (representing Paths,
AppNames, URLs, etc.). They don't necessarily have to be in the `legendary`
scope, should we increase our usage of Zod in the future