Adapt backend function to accept LegendaryCommand
As mentioned in the previous commit, Legendary commands are now objects.
Representing this required small adaptations basically everywhere, along with
a new function, "commandToArgsArray". This function converts a "new" command
object to the "old" `string[]` representation, which can then be passed to
`callRunner`/used elsewhere.

There are a few rather unsightly things here, most of which stem from the fact
that backwards compatibility has to be kept. Assuming this approach works out,
we can look into applying Zod in other areas of the codebase as well.