[Tech] Type-check Legendary commands (#2918)
* Don't require a runRunnerCommand in a LibraryManager

This wasn't actually used anywhere, and upcoming commits will make Legendary's
runRunnerCommand incompatible with this signature

* Add type definitions for all Legendary commands used in the project

This allows for stricter type checking (coming in the next commit).
The general idea behind this is:
- each command is represented by an object
- each of those objects has a `subcommand` property containing the name of the
  command, and other required properties for required parameters (for
  example, `launch` requires an `AppName`)
- optional parameters are represented by optional object properties with the
  only valid value of `true`

Base types (seen in `base.ts`) are general-purpose types (representing Paths,
AppNames, URLs, etc.). They don't necessarily have to be in the `legendary`
scope, should we increase our usage of Zod in the future

* Adapt backend function to accept LegendaryCommand

As mentioned in the previous commit, Legendary commands are now objects.
Representing this required small adaptations basically everywhere, along with
a new function, "commandToArgsArray". This function converts a "new" command
object to the "old" `string[]` representation, which can then be passed to
`callRunner`/used elsewhere.

There are a few rather unsightly things here, most of which stem from the fact
that backwards compatibility has to be kept. Assuming this approach works out,
we can look into applying Zod in other areas of the codebase as well.

* Split `getWineFlags`'s functionality into two functions

* Move `commandToArgsArray` into legendary/library