backend/nile: handle update requests for empty installed game list gracefully.
If no game is installed, nile list-updates will return an empty string.
Unfortunately, this is not valid JSON and JSON.parse() later bails out
with an error.

We're good to go if no game is installed either way, so break out early
and don't try to parse an empty string.