[Fix] Several System Information improvements (#3070)
* Report Host OS info in a Flatpak container

On Flatpak, `/run/host/os-release` will exist & contain the host OS'
information. I intended to report this information, but missed a
`break` in the loop (so it'd always find `/etc/os-release`
afterwards and read that out instead)

* Pad PCI IDs with 0

This avoids issues with very old NVIDIA GPUs.
The GPU in question was a Quadro 2000, with an ID of "0dd8". Heroic
searched for just "dd8", which is not in the PCI IDs list

* Make the memory detection more resistant to locale issues

- Certain non-english characters aren't considered "Word characters" in
  Regex, so \S (any non-whitespace) is now accepted
- The `free` command is ran with `LANG=C` to force English output even
  on non-english systems
- If the match fails for any other reason, we simply return 0 bytes
  total/used (which is at least a valid number, even if it doesn't make
  sense)

* Catch `getSystemInfo`-related errors

When initializing Heroic's general / the game's log, we don't *have to*
get system information for anything else to work; we can just report
that there was an error & carry on