backend/launcher: don't try to pass empty string as additional argument if falsy (empty, null, ...)
If the additional command line arguments are unset/set to an empty
string, the sideload command launcher likewise used to pass an empty
argument to the launched application.

While that is usually harmless, some applications try to parse this
argument and (rightfully) complain and usually exit with an error.

Make sure that we only pass additional arguments if they are actually
set, fixing that issue.

This has been encountered with Palia's launcher.

While at it, also fix a log line by including additional arguments and a
typo in a comment.