Add logo support in GamePage and GamePicture
In GamePage component, add logo variable and pass it to GamePicture component.
In GamePicture component, add logo parameter; add logo element (same as
the CachedImage logo element in GameCard)

There was no need to add CSS, since there was  already CSS for logo in
GamePicture  that positions it at the bottom center.