Extract archives into their destination directories directly
Before, we told tar to extract into the tool's directory (not into the
final destination folder), and thus relied on the archive containing
a sub-directory with the name we were looking for. DXVK-NVAPI doesn't
do this, so we overwrote this behavior for it only.

Now, we create the destination folder and extract to it, and use the
`strip` parameter for tools that ship inside a sub-directory to remove
this directory.

This makes sure that the destination directory always exists, which
resolves issues installing DXVK-NVAPI.