[Feat] Support Epic's launch-able addons/DLCs (#3317)
* Misc: Consolidate all InstallInfo into one type

* Legendary: Accept an AppName in `loadFile`

This also factors out the actual reading & parsing of the file into a
new function (which will be required by the next commit)

* Don't add the "--skip-version-check" flag in the Frontend

This de-couples the shape of "launchArguments" from this option, which
allows us to encode more detailed information in it (which will happen
in the next commit)

* Support Epic's DLC launch options

The backend is now responsible for providing launch options. For GOG
this required no changes, Amazon and sideloaded Games don't support
launch options at all (so they just return an empty array).
Legendary was a little more difficult: Other than the options listed in
the game's metadata, launch options can also be added by the following:
- If a DLC has an executable set, that DLC is also launch-able through
  Legendary
- If a DLC has the "AdditionalCommandLine" attribute set, launching the
  game with the parameters specified is supposed to launch the DLC
The latter is rather easy to implement, while the former requires
re-thinking how we do launch arguments:

"LaunchOption" is now an object with a type & type-specific properties.
Its former shape is preserved in "BaseLaunchOption", which is still the
only available option for anything non-Legendary.
"DLCLaunchOption" is new, and holds the AppName of a DLC to launch
instead of the main game.

With this done, Legendary's "launch" function can simply replace the
AppName passed to Legendary if a DLC launch option is passed to it.