Pass `--skip-sdl` when updating/repairing Epic games
If we're updating/repairing a game with SDL that was either synced from
EGS or imported (or is missing SDL tags for some other reason), the
command will get stuck waiting for the user to choose these tags.

Passing `--skip-sdl` resolves this issue. It'll always set the "default"
tags, which might not be the user's choice, but it's better than getting
stuck like this.