[Fix] Sideloaded games on Windows (#3562)
* Support empty commandParts in callRunner's Windows case

In case the array is empty, we have to make sure to not pass the parameter at
all (otherwise PS will error out)

This is used by sideloaded games

* Correctly pass the "runner" path for sideloaded games

Sideloaded games pass their executable path as the runner. Before, this was done
 a little incorrectly. Assume the selected executable is
`C:\Windows\System32\cmd.exe`. Before, the runner would then be:
{
  bin: `C:\Windows\System32\cmd.exe`
  dir: `C:\Windows\System32\`
}

callRunner then just `join`s together these paths, resulting in
`C:\Windows\System32\C:\Windows\System32\cmd.exe`. This is obviously wrong and
will not work.

Now, we correctly pass just the bin for `bin` (`cmd.exe` in our example). This
is also in-line with how regular runners work

* Add "./" in callRunner directly

This was somewhat flawed before; callRunner relied on an implementation detail
of `splitPathAndName` (it adding "./" to the "bin"). The relevant code was now
moved to callRunner itself