Fix game settings showing wrong value after changing global config
The Frontend is a little bit naive with its "cache" of game config
values: If a GameConfig is stored, it blindly uses the values from that
stored object. This is fine for the most part, but not in the following
case:
1. User opens game settings dialog (thus caching a full GameConfig
   object with default values for some keys)
2. User navigates to the global settings and changes something applying
   to both global and game config (this updates the *Global*Config, but
   not any GameConfigs)
3. User returns to game settings screen, the system will still have the
   old values cached. One of those values might be the old default
   value the user just changed, which is now erroneously displayed

The quick fix for this is just to throw away all GameConfigs when the
GlobalConfig changes. This isn't ideal, but I doubt it's going to matter