Migrate `libraryStatus` and make it a Record
This was definitely the biggest change yet. Most of it was just taking a
`runner` everywhere (to build a key to index into the record)

This still has one issue (the library isn't refreshed after a game
install/update finishes), it'll be resolved by future commits. Still, this
should be a noticeable speed boost