Implement adding & removing games from storage libraries
This is only really a sketch for now, these functions aren't used yet

GameManagers now have a property specifying whether they support "storage
libraries" (name used to differentiate "libraries on a user account" from
"libraries on the hard drive"). If they report support, they have to implement
- importFromStorageLibrary, used to import games when adding an existing library
- removeFromStorageLibrary, used to remove a game from the manager's internal
  library without removing game files
In the future, `install`, `update`, and so on will also receive an optional
`library` parameter. GameManagers are then expected to modify the entry for the
game in the library (required methods will be implemented in the next commits)