[chore][tech] Update dependencies that do not require significant API changes.
Notable updates that caused a bit of code churn are eslint and prettier.

The new eslint and especially its react rules do a bunch of checks around manual memoization because the React Compiler cannot do anything with those annotations - I've temporarily set a couple of the new options to Warnings until I can actually evaluate the performance impact of React Compiler to see if the code changes are worth it.